requirements.txt
llama-index>=0.7.27
groq-llm>=0.1.0
sentence-transformers
faiss-cpu
gradio
python-dotenv
llama-index>=0.7.27
sentence-transformers
faiss-cpu
gradio
python-dotenv
